+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGELOG
 
+### 1.0.1 (unreleased)
+
+- Optimize index size by removing ranges that have codepoints embedded
+- Optimize index size by substituting common words
+- Fix missing Tangut ideographs
+
 ### 1.0.0
 
 - Initial release